Change PathFinding + fix alloc memory

+++ headers/abstract/multipleobject.h
@@ -6,9 +6,9 @@
 class MultipleObject{
 
 public:
-    virtual bool add(int x, int y) =0;
-    virtual bool remove(int x, int y) =0; 
-    virtual bool has(int x, int y) =0;
+    virtual bool add(int8_t x, int8_t y) =0;
+    virtual bool remove(int8_t x, int8_t y) =0; 
+    virtual bool has(int8_t x, int8_t y) =0;
     virtual bool clear() =0;
     virtual std::string toString() =0;
 };
