+++ headers/manager/arrive.h
@@ -11,11 +11,11 @@ class Arrive : public SimpleObject{
 
 public:
     static Arrive * getInstance();
-    bool setPosition(int x, int y) override{
+    bool setPosition(int8_t x, int8_t y) override{
         position->first = x;
         position->second = y;
     }
-    pair<int, int> getPosition() override{
+    pair<int8_t, int8_t> getPosition() override{
         return *position;
     }
     string toString() override;
@@ -28,9 +28,9 @@ public:
     
 private:
     static Arrive * instance;
-    pair<int, int> *position;
+    pair<int8_t, int8_t> *position;
     Arrive(){
-        position = new pair<int, int>();
+        position = new pair<int8_t, int8_t>();
         position->first = -1;
         position->second = -1;
     }
