+++ headers/manager/car.h
@@ -12,11 +12,11 @@ class Car : public SimpleObject{
 public:
     static Car * getInstance();
     void sendCommand(); //send command to car from position X & Y
-    bool setPosition(int x, int y) override{
+    bool setPosition(int8_t x, int8_t y) override{
         position->first = x;
         position->second = y;
     }
-    pair<int, int> getPosition() override{
+    pair<int8_t, int8_t> getPosition() override{
         return *position;
     }
     string toString() override;
@@ -28,9 +28,9 @@ public:
     }
 private:
     static Car * instance;
-    pair<int, int> *position;
+    pair<int8_t, int8_t> *position;
     Car(){
-        position = new pair<int, int>();
+        position = new pair<int8_t, int8_t>();
         position->first = -1;
         position->second = -1;
     }
